part_4/content_assistant_server/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "content_assistant"
authors = [{ name = "Markell Richards" }]
version = "0.0.1"
requires-python = ">=3.12"
readme = "README.md"

dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "tenacity",
    "asyncio",
    "llama-index-core",
    "llama-index-llms-openai",
    "tavily-python",
    "openai",
    "pydantic-settings",
    "minio",
    "JSON-log-formatter",
    "sqlalchemy",
    "alembic",
    "psycopg2-binary",
    "boto3",
    "asyncpg",
]

[project.optional-dependencies]
dev = ["ruff", "black"]

# [tool.alembic]
